Extract Bazel Dependencies target creation
Code around creating the Bazel Dependencies target was taking up a lot of what was in `Generator+AddTargets.swift`. This will allow adding specific tests for it in the future as well.
