+++ test/fixtures/generator/bwb_spec.json
@@ -426,6 +426,7 @@
                     "tools/generator/src/Errors.swift",
                     "tools/generator/src/FilePath.swift",
                     "tools/generator/src/FilePathResolver.swift",
+                    "tools/generator/src/Generator+AddBazelDependenciesTarget.swift",
                     "tools/generator/src/Generator+AddTargets.swift",
                     "tools/generator/src/Generator+CreateFilesAndGroups.swift",
                     "tools/generator/src/Generator+CreateProducts.swift",
