+++ tools/generator/src/Generator+AddTargets.swift
@@ -2,87 +2,15 @@ import PathKit
 import XcodeProj
 
 extension Generator {
-    // Xcode likes this list as a string, and apparently in reverse
-    static let allPlatforms = """
-watchsimulator \
-watchos \
-macosx \
-iphonesimulator \
-iphoneos \
-driverkit \
-appletvsimulator \
-appletvos
-"""
-
-    static let bazelExec = #"""
-env -i \
-  DEVELOPER_DIR="$DEVELOPER_DIR" \
-  HOME="$HOME" \
-  PATH="${PATH//\/usr\/local\/bin//opt/homebrew/bin:/usr/local/bin}" \
-  USER="$USER" \
-  "$BAZEL_PATH"
-"""#
-
-    static let setup = #"""
-if [ "$ACTION" == "indexbuild" ]; then
-  # We use a different output base for Index Build to prevent normal builds and
-  # indexing waiting on bazel locks from the other
-  output_base="$OBJROOT/bazel_output_base"
-fi
-
-output_path=$(\#(bazelExec) \
-  ${output_base:+--output_base "$output_base"} \
-  info \
-  --experimental_convenience_symlinks=ignore \
-  output_path)
-external="${output_path%/*/*/*}/external"
-
-# We only want to modify `$LINKS_DIR` during normal builds since Indexing can
-# run concurrent to normal builds
-if [ "$ACTION" != "indexbuild" ]; then
-  mkdir -p "$LINKS_DIR"
-  cd "$LINKS_DIR"
-
-  # Add BUILD and DONT_FOLLOW_SYMLINKS_WHEN_TRAVERSING_THIS_DIRECTORY_VIA_A_RECURSIVE_TARGET_PATTERN
-  # files to the internal links directory to prevent Bazel from recursing into
-  # it, and thus following the `external` symlink
-  touch BUILD
-  touch DONT_FOLLOW_SYMLINKS_WHEN_TRAVERSING_THIS_DIRECTORY_VIA_A_RECURSIVE_TARGET_PATTERN
-
-  # Need to remove the directories that Xcode creates as part of output prep
-  rm -rf external
-  rm -rf gen_dir
-
-  ln -sf "$external" external
-  ln -sf "$BUILD_DIR/bazel-out" gen_dir
-fi
-
-cd "$BUILD_DIR"
-
-rm -rf external
-rm -rf real-bazel-out
-
-ln -sf "$external" external
-ln -sf "$output_path" real-bazel-out
-ln -sfn "$PROJECT_DIR" SRCROOT
-"""#
-
     static func addTargets(
         in pbxProj: PBXProj,
         for disambiguatedTargets: [TargetID: DisambiguatedTarget],
         products: Products,
         files: [FilePath: File],
         filePathResolver: FilePathResolver,
-        xcodeprojBazelLabel: String
+        bazelDependenciesTarget: PBXAggregateTarget
     ) throws -> [TargetID: PBXNativeTarget] {
         let pbxProject = pbxProj.rootObject!
-
-        let bazelDependenciesTarget = try createBazelDependenciesTarget(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver,
-            xcodeprojBazelLabel: xcodeprojBazelLabel
-        )
 
         let sortedDisambiguatedTargets = disambiguatedTargets
             .sortedLocalizedStandard(\.value.name)
@@ -157,319 +85,6 @@ Product for target "\(id)" not found in `products`
         }
 
         return pbxTargets
-    }
-
-    private static func createBazelDependenciesTarget(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver,
-        xcodeprojBazelLabel: String
-    ) throws -> PBXAggregateTarget {
-        let pbxProject = pbxProj.rootObject!
-
-        let debugConfiguration = XCBuildConfiguration(
-            name: "Debug",
-            buildSettings: [
-                "ALLOW_TARGET_PLATFORM_SPECIALIZATION": true,
-                "BAZEL_PACKAGE_BIN_DIR": "rules_xcodeproj",
-                "INDEX_FORCE_SCRIPT_EXECUTION": true,
-                "SUPPORTED_PLATFORMS": allPlatforms,
-                "SUPPORTS_MACCATALYST": true,
-                "TARGET_NAME": "BazelDependencies",
-            ]
-        )
-        pbxProj.add(object: debugConfiguration)
-        let configurationList = XCConfigurationList(
-            buildConfigurations: [debugConfiguration],
-            defaultConfigurationName: debugConfiguration.name
-        )
-        pbxProj.add(object: configurationList)
-
-        let fetchExternalReposScript = createFetchExternalReposScript(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver,
-            xcodeprojBazelLabel: xcodeprojBazelLabel
-        )
-
-        let generateFilesScript = createGenerateFilesScript(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver,
-            xcodeprojBazelLabel: xcodeprojBazelLabel
-        )
-
-        let copyFilesScript = createCopyFilesScript(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver
-        )
-
-        let fixModuleMapsScript = createFixModulemapsScript(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver
-        )
-
-        let fixInfoPlistsScript = createFixInfoPlistsScript(
-            in: pbxProj,
-            files: files,
-            filePathResolver: filePathResolver
-        )
-
-        let pbxTarget = PBXAggregateTarget(
-            name: "Bazel Dependencies",
-            buildConfigurationList: configurationList,
-            buildPhases: [
-                fetchExternalReposScript,
-                generateFilesScript,
-                copyFilesScript,
-                fixModuleMapsScript,
-                fixInfoPlistsScript,
-            ].compactMap { $0 },
-            productName: "Bazel Dependencies"
-        )
-        pbxProj.add(object: pbxTarget)
-        pbxProject.targets.append(pbxTarget)
-
-        let attributes: [String: Any] = [
-            // TODO: Generate this value
-            "CreatedOnToolsVersion": "13.2.1",
-        ]
-        pbxProject.setTargetAttributes(attributes, target: pbxTarget)
-
-        return pbxTarget
-    }
-
-    private static func createFetchExternalReposScript(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver,
-        xcodeprojBazelLabel: String
-    ) -> PBXShellScriptBuildPhase? {
-        guard !files.containsGeneratedFiles else {
-            return nil
-        }
-
-        let script = PBXShellScriptBuildPhase(
-            name: "Fetch External Repositories",
-            outputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(externalFileListPath))
-                    .string,
-            ],
-            shellScript: #"""
-set -eu
-
-\#(setup)
-
-cd "$SRCROOT"
-
-\#(bazelExec) \
-  ${output_base:+--output_base "$output_base"} \
-  build \
-  --nobuild \
-  --experimental_convenience_symlinks=ignore \
-  \#(xcodeprojBazelLabel)
-
-"""#,
-            showEnvVarsInLog: false,
-            alwaysOutOfDate: true
-        )
-        pbxProj.add(object: script)
-
-        return script
-    }
-
-    private static func createGenerateFilesScript(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver,
-        xcodeprojBazelLabel: String
-    ) -> PBXShellScriptBuildPhase? {
-        guard files.containsGeneratedFiles else {
-            return nil
-        }
-
-        let script = PBXShellScriptBuildPhase(
-            name: "Generate Files",
-            outputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(externalFileListPath))
-                    .string,
-                filePathResolver
-                    .resolve(.internal(generatedFileListPath))
-                    .string,
-            ],
-            shellScript: #"""
-set -eu
-
-\#(setup)
-
-# Create parent directories of generated files, so the project navigator works
-# better faster
-
-mkdir -p bazel-out
-cd bazel-out
-
-sed 's|\/[^\/]*$||' \
-  "\#(
-  filePathResolver
-      .resolve(.internal(rsyncFileListPath), useScriptVariables: true)
-      .string
-)" \
-  | uniq \
-  | while IFS= read -r dir
-do
-  mkdir -p "$dir"
-done
-
-cd "$SRCROOT"
-
-date +%s > "$INTERNAL_DIR/toplevel_cache_buster"
-
-\#(bazelExec) \
-  ${output_base:+--output_base "$output_base"} \
-  build \
-  --experimental_convenience_symlinks=ignore \
-  --output_groups=generated_inputs \
-  \#(xcodeprojBazelLabel)
-
-"""#,
-            showEnvVarsInLog: false,
-            alwaysOutOfDate: true
-        )
-        pbxProj.add(object: script)
-
-        return script
-    }
-
-    private static func createCopyFilesScript(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver
-    ) -> PBXShellScriptBuildPhase? {
-        guard files.containsGeneratedFiles else {
-            return nil
-        }
-
-        let script = PBXShellScriptBuildPhase(
-            name: "Copy Files",
-            inputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(generatedFileListPath))
-                    .string,
-            ],
-            outputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(copiedGeneratedFileListPath))
-                    .string,
-            ],
-            shellScript: #"""
-set -eu
-
-cd "$BAZEL_OUT"
-
-# Sync to "$BUILD_DIR/bazel-out". This is the same as "$GEN_DIR" for normal
-# builds, but is different for Index Builds. `PBXBuildFile`s will use the
-# "$GEN_DIR" version, so indexing might get messed up until they are normally
-# generated. It's the best we can do though as we need to use the `gen_dir`
-# symlink.
-rsync \
-  --files-from "\#(
-    filePathResolver
-        .resolve(.internal(rsyncFileListPath), useScriptVariables: true)
-        .string
-)" \
-  --chmod=u+w \
-  -L \
-  . \
-  "$BUILD_DIR/bazel-out"
-
-"""#,
-            showEnvVarsInLog: false
-        )
-        pbxProj.add(object: script)
-
-        return script
-    }
-
-    private static func createFixModulemapsScript(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver
-    ) -> PBXShellScriptBuildPhase? {
-        guard files.containsModulemaps else {
-            return nil
-        }
-
-        let script = PBXShellScriptBuildPhase(
-            name: "Fix Modulemaps",
-            inputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(modulemapsFileListPath))
-                    .string,
-            ],
-            outputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(fixedModulemapsFileListPath))
-                    .string,
-            ],
-            shellScript: #"""
-set -eu
-
-while IFS= read -r input; do
-  output="${input%.modulemap}.xcode.modulemap"
-  perl -p -e \
-    's%^(\s*(\w+ )?header )(?!("\.\.(\/\.\.)*\/|")(bazel-out|external)\/)("(\.\.\/)*)(.*")%\1\6SRCROOT/\8%' \
-    < "$input" \
-    > "$output"
-done < "$SCRIPT_INPUT_FILE_LIST_0"
-
-"""#,
-            showEnvVarsInLog: false
-        )
-        pbxProj.add(object: script)
-
-        return script
-    }
-
-    private static func createFixInfoPlistsScript(
-        in pbxProj: PBXProj,
-        files: [FilePath: File],
-        filePathResolver: FilePathResolver
-    ) -> PBXShellScriptBuildPhase? {
-        guard files.containsInfoPlists else {
-            return nil
-        }
-
-        let script = PBXShellScriptBuildPhase(
-            name: "Fix Info.plists",
-            inputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(infoPlistsFileListPath))
-                    .string,
-            ],
-            outputFileListPaths: [
-                filePathResolver
-                    .resolve(.internal(fixedInfoPlistsFileListPath))
-                    .string,
-            ],
-            shellScript: #"""
-set -eu
-
-while IFS= read -r input; do
-  output="${input%.plist}.xcode.plist"
-  cp "$input" "$output"
-  plutil -remove UIDeviceFamily "$output" || true
-done < "$SCRIPT_INPUT_FILE_LIST_0"
-
-"""#,
-            showEnvVarsInLog: false
-        )
-        pbxProj.add(object: script)
-
-        return script
     }
 
     private static func createHeadersPhase(
@@ -774,28 +389,6 @@ private struct SourceFile: Hashable {
     }
 }
 
-extension Inputs {
-    var containsSourceFiles: Bool {
-        return !(srcs.isEmpty && nonArcSrcs.isEmpty)
-    }
-}
-
-extension Dictionary where Key == FilePath {
-    var containsModulemaps: Bool {
-        contains(where: { filePath, _ in
-            return filePath.type == .generated
-                && filePath.path.extension == "modulemap"
-        })
-    }
-
-    var containsInfoPlists: Bool {
-        contains(where: { filePath, _ in
-            return filePath.type == .generated
-                && filePath.path.lastComponent == "Info.plist"
-        })
-    }
-}
-
 private extension Path {
     var isHeader: Bool {
         switch `extension` {
@@ -804,29 +397,5 @@ private extension Path {
         case "hpp": return true
         default: return false
         }
-    }
-}
-
-// Re-implementation of helper on `PBXNativeTarget` until our patch lands:
-// https://github.com/tuist/XcodeProj/pull/677
-extension PBXAggregateTarget {
-    /// Adds a local target dependency to the target.
-    ///
-    /// - Parameter target: dependency target.
-    /// - Returns: target dependency reference.
-    /// - Throws: an error if the dependency cannot be created.
-    func addDependency(target: PBXTarget, in pbxProj: PBXProj) throws -> PBXTargetDependency? {
-        let pbxProject = pbxProj.rootObject!
-        let proxy = PBXContainerItemProxy(containerPortal: .project(pbxProject),
-                                          remoteGlobalID: .object(target),
-                                          proxyType: .nativeTarget,
-                                          remoteInfo: target.name)
-        pbxProj.add(object: proxy)
-        let targetDependency = PBXTargetDependency(name: target.name,
-                                                   target: target,
-                                                   targetProxy: proxy)
-        pbxProj.add(object: targetDependency)
-        dependencies.append(targetDependency)
-        return targetDependency
     }
 }
